chore: move schema into the audit model
